% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_algorithmia.R
\name{GEKS_algorithmia}
\alias{GEKS_algorithmia}
\title{Algorithmia Wrapper for GEKS}
\usage{
GEKS_algorithmia(input_json)
}
\arguments{
\item{input_json}{an input json. See ?GEKS for the details of the JSON}
}
\value{
see ?GEKS for return details. This function will retun the exact same
results, except in json format rather than a list
}
\description{
Algorithmia Wrapper for GEKS
}
