% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{length_calc}
\alias{length_calc}
\title{Help calculate how many periods you need for desired index length}
\usage{
length_calc(index_start, window_length, verbose = TRUE)
}
\arguments{
\item{index_start}{how many periods back do you want the index to start (numeric)}

\item{window_length}{window length used}
}
\value{
The numeric representation of how many periods you will need in your input
dataset to calculate an index of length 'index_start' with a window length of 'window_length'
}
\description{
The FEWS function truncates the results by removing the first window. This
function helps you determine the amount of periods required to produced
the desired FEWS index output length
}
\examples{
#I want to calculate an index from June to June for a monthly index with a window length
#of 27 months

#index_start will be 13 as this is the number of periods for June to June

length_calc(index_start = 13, window_length = 27)

#Total periods required:  39

}
